refactor: refactor rvps as a mod
Refactor RVPS into a submodule of AS, which helps to
collect all the cgo code into one build.rs file to
prevent errors when linking.
