--- rvps/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "rvps"
-version = "0.1.0"
-edition = "2021"
-description = "reference-value-provider-service"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow = "1.0.57"
-serde_json = "1.0.81"
-serde = { version = "1.0.137", features = [ "derive" ] }
-redis = "0.21.5"
-chrono = { version = "0.4.19", features = [ "serde" ] }
-tempfile = "3.3.0"
-base64 = "0.13.0"
-log = "0.4.17"
-sled = "0.34.7"
-strum = "0.24.0"
-strum_macros = "0.24.0"
-
-[dev-dependencies]
-testing_logger = "0.1.1"
-walkdir = "2.3.2"
-serial_test = "0.9.0"
-sha2 = "0.10.2"
